update code p5

+++ BankingApp_winform/RecurringPaymentServices.h
@@ -86,7 +86,7 @@ ref class RecurringPaymentServices {
                                    L"hãy nộp thêm tiền vào tài khoản để tiến hành giao dịch";
 
                 NotificationsServices::InsertNotification(recurringPayments->UserAccountNumber, content);
-                return;
+                throw gcnew Exception(L"Khách hàng không đủ số dư để thanh toán định kì hóa đơn này");
             }
 
             // update code detail
@@ -114,9 +114,76 @@ ref class RecurringPaymentServices {
         }
     }
 
+    static void RegisterRecurringPayment(String^ customerCodeString, String^ accNum) {
+        try {
+            if (customerCodeString == "" || accNum == "") {
+                throw gcnew Exception(L"Vui lòng nhập đầy đủ thông tin");
+            }
+            
+            // kiểm tra mã dịch vụ đã tồn tại chưa
+            CustomerCodes ^ customerCode = CustomerCodesRepository::FindCustomerCodeByCodeString(customerCodeString);
+            if (customerCode == nullptr) {
+                throw gcnew Exception(L"Mã dịch vụ không tồn tại");
+            } else if (customerCode->Status == 0) {
+                throw gcnew Exception(L"Mã dịch vụ không còn hoạt động");
+            }
+
+            // kiểm tra  mã thanh toán đã được đăng kí chưa
+            RecurringPayments ^ recurringPayments = RecurringPaymentServices::GetRecurringPaymentByCustomerCodeId(customerCode->Id);
+            if (recurringPayments != nullptr) {
+                throw gcnew Exception(L"Mã khách hàng này đã được đăng kí định kì rồi");
+            }
+
+            // kiểm tra khách hàng đã đăng kí mã nào nào công ty này chưa
+            array<RecurringPayments ^> ^ myRecurringPayments = RecurringPaymentServices::GetRecurringPaymentsByUserAccNumber(accNum);
+            if (myRecurringPayments != nullptr) {
+                for (int i = 0; i < myRecurringPayments->Length; i++) {
+                    if (myRecurringPayments[i]->CustomerCodeId == customerCode->Id) {
+                        throw gcnew Exception(L"Mã khách hàng này đã được đăng kí định kì rồi");
+                    }
+                }
+            }
+
+            String ^ id = Utils::createUniqueID("RP");
+            RecurringPayments ^ newRecurringPayment =
+                gcnew RecurringPayments(id, accNum, customerCode->Id, Convert::ToInt32(DateTime::Now.Day));
+
+            RecurringPaymentsRepository::InsertOne(newRecurringPayment);
+
+            // thêm thông báo cho khách hàng và công ty
+            String ^ content = L"Đăng kí thanh toán định kì cho mã khách hàng " + customerCode->Code + L" thành công";
+            NotificationsServices::InsertNotification(accNum, content);
+
+            content = L"Khách hàng " + accNum + L" đã đăng kí thanh toán định kì cho mã khách hàng " + customerCode->Code;
+            NotificationsServices::InsertNotification(customerCode->CompanyAccountNumber, content);
+
+        } catch (Exception ^ ex) {
+            throw ex;
+        }
+    }
+
     static void DeleteRecurringPaymentById(String ^ id) {
         try {
+
+            // thêm thông báo cho khách hàng và công ty
+            RecurringPayments ^ recurringPayment = RecurringPaymentsRepository::GetOneById(id);
+            if (recurringPayment == nullptr) {
+                throw gcnew Exception(L"Không tìm thấy thông tin mã thanh toán định kì");
+            }
+
+            CustomerCodes ^ customerCode = CustomerCodesRepository::FindCustomerCodeById(recurringPayment->CustomerCodeId);
+            if (customerCode == nullptr) {
+                throw gcnew Exception(L"Không tìm thấy thông tin mã khách hàng");
+            }
+
+            String ^ content = L"Đã hủy thanh toán định kì cho mã khách hàng " + customerCode->Code;
+            NotificationsServices::InsertNotification(recurringPayment->UserAccountNumber, content);
+
+            content = L"Khách hàng " + recurringPayment->UserAccountNumber + L" đã hủy thanh toán định kì cho mã khách hàng " + customerCode->Code;
+            NotificationsServices::InsertNotification(customerCode->CompanyAccountNumber, content);
+
             RecurringPaymentsRepository::DeleteById(id);
+
         } catch (Exception ^ ex) {
             throw ex;
         }
